Refactor: 파일명 & 폴더명 수정 (#146)
