+++ .storybook/storybook.requires.js
@@ -68,10 +68,10 @@ const getStories = () => {
     "./modules/post/components/detail/duration/Duration.stories.tsx": require("../modules/post/components/detail/duration/Duration.stories.tsx"),
     "./modules/post/components/detail/duration/schedule/Schedule.stories.tsx": require("../modules/post/components/detail/duration/schedule/Schedule.stories.tsx"),
     "./modules/post/components/detail/duration/timePickerButton/TimePickerButton.stories.tsx": require("../modules/post/components/detail/duration/timePickerButton/TimePickerButton.stories.tsx"),
-    "./modules/post/components/detail/participant/member/Member.stories.tsx": require("../modules/post/components/detail/participant/member/Member.stories.tsx"),
-    "./modules/post/components/detail/participant/memberCount/MemberCount.stories.tsx": require("../modules/post/components/detail/participant/memberCount/MemberCount.stories.tsx"),
-    "./modules/post/components/detail/participant/members/Members.stories.tsx": require("../modules/post/components/detail/participant/members/Members.stories.tsx"),
-    "./modules/post/components/detail/participant/ParticipantCard.stories.tsx": require("../modules/post/components/detail/participant/ParticipantCard.stories.tsx"),
+    "./modules/post/components/detail/participants/member/Member.stories.tsx": require("../modules/post/components/detail/participants/member/Member.stories.tsx"),
+    "./modules/post/components/detail/participants/memberCount/MemberCount.stories.tsx": require("../modules/post/components/detail/participants/memberCount/MemberCount.stories.tsx"),
+    "./modules/post/components/detail/participants/members/Members.stories.tsx": require("../modules/post/components/detail/participants/members/Members.stories.tsx"),
+    "./modules/post/components/detail/participants/Participants.stories.tsx": require("../modules/post/components/detail/participants/Participants.stories.tsx"),
     "./modules/post/components/search/searchAndCreate/SearchAndCreateBar.stories.tsx": require("../modules/post/components/search/searchAndCreate/SearchAndCreateBar.stories.tsx"),
     "./modules/post/components/search/searchBar/SearchBar.stories.tsx": require("../modules/post/components/search/searchBar/SearchBar.stories.tsx"),
     "./modules/post/screens/MeetingDatePicker.stories.tsx": require("../modules/post/screens/MeetingDatePicker.stories.tsx"),
