+++ modules/post/components/detail/participants/Participants.spec.tsx
@@ -3,9 +3,9 @@ import { render, screen } from '@testing-library/react-native';
 
 import mockMembers from '@post/mocks/members';
 import { wrapper } from '@shared/components/ThemeProvider';
-import ParticipantCard from './ParticipantCard';
+import ParticipantCard from './Participants';
 
-describe('ParticipantCard Compoent', () => {
+describe('Participants Compoent', () => {
   test('멤버와 관련된 정보가 올바르게 렌더링 되어야 한다.', () => {
     render(<ParticipantCard users={mockMembers} />, wrapper);
