+++ modules/post/components/detail/participants/member/Member.stories.tsx
@@ -7,7 +7,7 @@ import Component from './Member';
 type Meta = ComponentMeta<typeof Component>;
 
 export default {
-  title: 'Detail - Participant',
+  title: 'Detail - Participants',
   component: Component,
   decorators: [
     (Story) => (
