+++ modules/post/components/detail/participants/member/style.ts
