+++ modules/post/components/detail/participants/type.tsx
